# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API manage many products of any company.",
        "title": "API GO",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080/api/v1",
    "paths": {
...
        "/localities": {
            "get": {
                "description": "Get all the localities available or an error if the list is empty or an internal error occurs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain.Locality"
                ],
                "summary": "Get all the localities available.",
                "responses": {
                    "200": {
                        "description": "List of all localities",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Locality"
                            }
                        }
                    },
                    "404": {
                        "description": "Localities not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new locality or an error if that locality cannot be created or an internal error occurs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain.Locality"
                ],
                "summary": "Create a new locality or an error if that locality cannot be created.",
                "parameters": [
                    {
                        "description": "Struct of Locality domain",
                        "name": "Locality",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Locality"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New Locality created",
                        "schema": {
                            "$ref": "#/definitions/domain.Locality"
                        }
                    },
                    "409": {
                        "description": "locality already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "invalid or missing field",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/localities/reportSellers": {
            "get": {
                "description": "Get a report of sellers by locality using its id. Or show all the sellers of all the localities if the id is not provided.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain.Locality"
                ],
                "summary": "Get a report of sellers by locality.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "locality id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report of sellers by locality",
                        "schema": {
                            "$ref": "#/definitions/domain.ReportSellers"
                        }
                    },
                    "400": {
                        "description": "error getting the report for the requested ID. Id must be greater than 0",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "sellers not found for the requested ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "server internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/locality/{id}": {
            "get": {
                "description": "Get a locality by id or an error if that id not exists or an internal error occurs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain.Locality"
                ],
                "summary": "Get a locality by id or an error if that id not exists.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id from the locality",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Locality requested",
                        "schema": {
                            "$ref": "#/definitions/domain.Locality"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Locality not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Locality": {
            "type": "object",
            "properties": {
                "country_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "locality_name": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "integer"
                },
                "province_name": {
                    "type": "string"
                }
            }
        },
...
        "domain.ReportSellers": {
            "type": "object",
            "properties": {
                "locality_id": {
                    "type": "integer"
                },
                "locality_name": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "integer"
                },
                "sellers_count": {
                    "type": "integer"
                }
            }
        },
...
    }
}